Fix classified editing in Heretic, add menu to create Pick and Classified items in profile
