+++ indra/newview/llpanelclassified.h
@@ -180,100 +180,4 @@ private:
     static panel_list_t sAllPanels;
 };
 
-class LLPanelClassifiedEdit : public LLPanelClassifiedInfo
-{
-    LOG_CLASS(LLPanelClassifiedEdit);
-public:
-
-    static LLPanelClassifiedEdit* create();
-
-    virtual ~LLPanelClassifiedEdit();
-
-    /*virtual*/ BOOL postBuild() override;
-
-    void fillIn(const LLSD& key);
-
-    /*virtual*/ void onOpen(const LLSD& key) override;
-
-    /*virtual*/ void processProperties(void* data, EAvatarProcessorType type) override;
-
-    /*virtual*/ BOOL isDirty() const override;
-
-    /*virtual*/ void resetDirty() override;
-
-    void setSaveCallback(const commit_signal_t::slot_type& cb);
-
-    void setCancelCallback(const commit_signal_t::slot_type& cb);
-
-    /*virtual*/ void resetControls() override;
-
-    bool isNew() { return mIsNew; }
-
-    bool isNewWithErrors() { return mIsNewWithErrors; }
-
-    bool canClose();
-
-    void draw() override;
-
-    void stretchSnapshot();
-
-    U32 getCategory();
-
-    void setCategory(U32 category);
-
-    U32 getContentType();
-
-    void setContentType(U32 content_type);
-
-    bool getAutoRenew();
-
-    S32 getPriceForListing();
-
-protected:
-
-    LLPanelClassifiedEdit();
-
-    void sendUpdate();
-
-    void enableVerbs(bool enable);
-
-    void enableEditing(bool enable);
-
-    void showEditing(bool show);
-
-    std::string makeClassifiedName();
-
-    void setPriceForListing(S32 price);
-
-    U8 getFlags();
-
-    std::string getLocationNotice();
-
-    bool isValidName();
-
-    void notifyInvalidName();
-
-    void onSetLocationClick();
-    void onChange();
-    void onSaveClick();
-
-    void doSave();
-
-    void onPublishFloaterPublishClicked();
-
-    void onTexturePickerMouseEnter(LLUICtrl* ctrl);
-    void onTexturePickerMouseLeave(LLUICtrl* ctrl);
-
-    void onTextureSelected();
-
-private:
-    bool mIsNew;
-    bool mIsNewWithErrors;
-    bool mCanClose;
-
-    LLFloaterPublishClassified* mPublishFloater;
-
-    commit_signal_t mSaveButtonClickedSignal;
-};
-
 #endif // LL_LLPANELCLASSIFIED_H
