+++ indra/newview/skins/default/xui/en/panel_edit_classified.xml
@@ -1,0 +1,320 @@
+<?xml version="1.0" encoding="utf-8" standalone="yes" ?>
+<panel
+ background_visible="true"
+ bevel_style="in"
+ follows="left|top|right|bottom"
+ height="569"
+ label="Edit Classified"
+ layout="topleft"
+ left="0"
+ min_height="350"
+ name="panel_edit_classified"
+ help_topic="profile_edit_classified"
+ top="0"
+ width="333">
+ <panel.string
+  name="location_notice">
+    (will update after save)
+ </panel.string>
+ <string name="publish_label">
+  Publish
+ </string>
+ <string name="save_label">
+  Save
+ </string>
+  <button
+     follows="top|right"
+     height="23"
+     image_overlay="BackArrow_Off"
+     layout="topleft"
+     name="back_btn"
+     picture_style="true"
+     left="12"
+     tab_stop="false"
+     top="2"
+     width="23" />
+   <text
+     type="string"
+     length="1"
+     follows="top"
+     font="SansSerifHugeBold"
+     height="26"
+     layout="topleft"
+     left_pad="10"
+     name="title"
+     text_color="LtGray"
+     top="0"
+     width="250">
+        Edit Classified
+   </text>
+   <scroll_container
+    follows="all"
+    height="502"
+    layout="topleft"
+    left="8"
+    top_pad="10"
+    name="profile_scroll"
+    reserve_scroll_corner="false"
+    opaque="true"
+    width="312">
+    <panel
+     name="scroll_content_panel"
+     follows="left|top"
+     min_height="300"
+     layout="topleft"
+     top="0"
+     background_visible="false"
+     height="690"
+     left="0"
+     width="285">
+      <panel
+         name="snapshot_panel"
+         layout="topleft"
+         follows="left|top|right"
+         height="197"
+         left="10"
+         top="10"
+         width="272">
+      <texture_picker
+       follows="left|top|right"
+       height="197"
+       width="272"
+       layout="topleft"
+       top="0"
+       left="0"
+       name="classified_snapshot" />
+       <icon
+           height="197"
+           image_name="spacer24.tga"
+           layout="topleft"
+           name="edit_icon"
+           label=""
+           tool_tip="Click to select an image"
+           top="0"
+           left="0"
+           width="272" />
+       </panel>
+        <text
+         type="string"
+         length="1"
+         follows="left|top"
+         height="15"
+         font="SansSerifSmall"
+         font.style="BOLD"
+         layout="topleft"
+         left="10"
+         top="215"
+         name="Name:"
+         text_color="white"
+         width="280">
+            Title:
+        </text>
+        <line_editor
+         follows="left|top|right"
+         font="SansSerif"
+         height="20"
+         layout="topleft"
+         left="10"
+         top_pad="2"
+         max_length="63"
+         name="classified_name"
+         prevalidate_callback="ascii"
+         text_color="black"
+         width="273" />
+        <text
+         type="string"
+         length="1"
+         follows="left|top"
+         height="15"
+         font="SansSerifSmall"
+         font.style="BOLD"
+         layout="topleft"
+         left="10"
+         top_pad="20"
+         name="description_label"
+         text_color="white"
+         width="280">
+            Description:
+        </text>
+        <text_editor
+         follows="left|top|right"
+         height="100"
+         width="273"
+         hide_scrollbar="false"
+         layout="topleft"
+         left="10"
+         top_pad="2"
+         max_length="1023"
+         name="classified_desc"
+         text_color="black"
+         word_wrap="true" />
+        <text
+         type="string"
+         length="1"
+         font="SansSerifSmall"
+         font.style="BOLD"
+         follows="left|top"
+         height="15"
+         layout="topleft"
+         left="10"
+         name="location_label"
+         text_color="white"
+         top_pad="20"
+         width="280">
+            Location:
+        </text>
+        <text
+         type="string"
+         length="1"
+         follows="left|top"
+         height="30"
+         layout="topleft"
+         left="10"
+         name="classified_location"
+         right="-10"
+         top_pad="2"
+         width="280"
+         word_wrap="true">
+            loading...
+        </text>
+        <button
+         follows="left|top"
+         height="23"
+         label="Set to Current Location"
+         layout="topleft"
+         left="10"
+         top_pad="5"
+         name="set_to_curr_location_btn"
+         width="200" />
+        <text
+         follows="left|top"
+         font.style="BOLD"
+         height="10"
+         layout="topleft"
+         left="10"
+         name="category_label"
+         text_color="white"
+         top_pad="15"
+         value="Category:"
+         width="250" />
+        <combo_box
+         follows="left|top" 
+         height="23" 
+         label=""
+	     left="10" 
+         name="category" 
+         top_pad="5"
+         width="156" />
+        <text
+         follows="left|top"
+         font.style="BOLD"
+         height="10"
+         layout="topleft"
+         left="10"
+         name="content_type_label"
+         text_color="white"
+         top_pad="15"
+         value="Content type:"
+         width="250" />
+        <icons_combo_box
+         follows="left|top"
+         height="23"
+         label="General Content"
+         layout="topleft"
+         left="10"
+         name="content_type"
+         top_pad="5"
+         width="156">
+            <icons_combo_box.drop_down_button
+             image_overlay="Parcel_PG_Light"
+             image_overlay_alignment="left"
+             imgoverlay_label_space="3"
+             pad_left="3"/>
+            <icons_combo_box.item
+             label="Moderate Content"
+             name="mature_ci"
+             value="Mature">
+                <item.columns
+                 halign="center"
+                 type="icon"
+                 value="Parcel_M_Light"
+                 width="20"/>
+            </icons_combo_box.item>
+            <icons_combo_box.item
+             label="General Content"
+             name="pg_ci"
+             value="PG">
+                <item.columns
+                 halign="center"
+                 type="icon"
+                 value="Parcel_PG_Light"
+                 width="20"/>
+            </icons_combo_box.item>
+        </icons_combo_box>
+        <text
+         follows="left|top"
+         font.style="BOLD"
+         height="10"
+         layout="topleft"
+         left="10"
+         name="price_for_listing_label"
+         text_color="white"
+         top_pad="15"
+         value="Price for listing:"
+         width="250" />
+        <spinner
+         decimal_digits="0"
+         follows="left|top"
+         halign="left"
+         height="23"
+         increment="1"
+         label_width="20"
+         label="L$"
+         v_pad="10"
+         layout="topleft"
+         left="10"
+         value="50"
+         min_val="50"
+         max_val="99999"
+         name="price_for_listing"
+         top_pad="5"
+         tool_tip="Price for listing."
+         width="105" />
+        <check_box
+         height="16"
+         label="Auto renew each week"
+         layout="topleft"
+         left="10"
+         name="auto_renew"
+         top_pad="15"
+         width="250" />
+     </panel>
+    </scroll_container>
+    <panel
+     follows="left|right|bottom"
+     height="23"
+     label="bottom_panel"
+     layout="topleft"
+     left="8"
+     name="bottom_panel"
+     top_pad="5"
+     width="303">
+        <button
+         follows="bottom|left"
+         height="23"
+         label="[LABEL]"
+         layout="topleft"
+         name="save_changes_btn"
+         left="0"
+         top="0"
+         width="152" />
+        <button
+         follows="bottom|left"
+         height="23"
+         label="Cancel"
+         layout="topleft"
+         name="cancel_btn"
+         left_pad="3"
+         width="153" />
+    </panel>
+</panel>
